agent/requirements.txt
# LangChain + Vector Database Dependencies
langchain>=0.1.0
langchain-anthropic>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0
